api cleanup

+++ src/buffer-struct.h
@@ -4,8 +4,8 @@
 
 struct medusa_buffer {
         void *buffer;
-        int length;
-        int size;
+        int64_t length;
+        int64_t size;
 };
 
 int medusa_buffer_init (struct medusa_buffer *buffer);
